+++ src/swagger/schemas/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Export all Swagger schemas
+ */
+
+import { USDQuote, BSVQuote, Quotes } from "./quotes";
+import { MarketData, SearchResult, SearchResponse } from "./market";
+import { StatusResponse } from "./status";
+
+export const schemas = {
+  USDQuote,
+  BSVQuote,
+  Quotes,
+  MarketData,
+  SearchResult,
+  SearchResponse,
+  StatusResponse,
+};